qc/cellranger: relocated to 'qc.apps.cellranger'.
